update Banking.java
Added the fields :

integer for the current balance

integer for the previous transaction amount

String for the name of the customer

String for the customer ID.
Added methods:

deposit method which passes through an integer representing the amount specified by the user.
Withdraw method which passes through an integer representing the amount specified by the user.
A method that returns information about the last transactions